+++ src/components/HomeButtons.tsx
@@ -1,8 +1,4 @@
 import React from 'react';
-import { Button } from '@material-ui/core';
-import { ButtonProps } from '@material-ui/core/Button';
-import { Link } from 'react-router-dom';
-import styled, { css } from 'styled-components';
 import {
   BananaIcon,
   BusIcon,
@@ -17,7 +13,9 @@ import {
   WifiIcon
 } from './Icons';
 import { THomeButtonAnchor, THomeButtonRouterLink } from '../types';
-import { colors, fonts } from '../App.styles';
+import { colors } from '../App.styles';
+import { HomeRouterLink, HomeAnchorLink } from './HomeLink';
+import HomeButton from './HomeButton';
 
 const routerLinkButtons: THomeButtonRouterLink[] = [
   {
@@ -109,87 +107,6 @@ const coordinatedEntryButton: THomeButtonAnchor = {
   color: colors.rosewood,
   target: '_blank'
 };
-
-interface HomeLinkPropsBase {
-  children: React.ReactElement | React.ReactElement[];
-  key: string;
-}
-
-type HomeRouterLinkProps = HomeLinkPropsBase & THomeButtonRouterLink;
-type HomeAnchorProps = HomeLinkPropsBase & THomeButtonAnchor;
-
-const HomeLinkStyles = css`
-  display: block;
-  flex: 1 1 50%;
-  padding: 4px;
-  text-decoration: none;
-  width: 100%;
-`;
-
-const HomeRouterLink = styled((props: HomeRouterLinkProps) => {
-  const { children, key, linkProps, ...rest } = props;
-  return (
-    <Link {...linkProps} key={key} {...rest}>
-      {children}
-    </Link>
-  );
-})`
-  ${HomeLinkStyles}
-`;
-
-const HomeAnchorLink = styled((props: HomeAnchorProps) => {
-  const { children, key, href, target, ...rest } = props;
-  return (
-    <a key={key} href={href} target={target} {...rest}>
-      {children}
-    </a>
-  );
-})`
-  ${HomeLinkStyles}
-`;
-interface HomeButtonProps extends ButtonProps {
-  readonly buttonColor: string;
-}
-
-const HomeButton = styled((props: HomeButtonProps) => {
-  const { buttonColor, ...rest } = props;
-  return <Button {...rest} />;
-})`
-  && {
-    align-items: stretch;
-    background: ${(props: HomeButtonProps) =>
-      props.buttonColor || colors.greyDark};
-    border-radius: 0;
-    color: ${colors.white};
-    display: flex;
-    height: 100%;
-    padding: 10px;
-    text-decoration: none;
-    width: 100%;
-  }
-  &&:hover {
-    background: ${(props: HomeButtonProps) =>
-      props.buttonColor || colors.greyDark};
-    filter: brightness(95%);
-  }
-  > span {
-    align-items: flex-start;
-    display: flex;
-    flex-direction: column;
-    font-family: ${fonts.openSans};
-    font-size: 22px;
-    font-weight: 700;
-    justify-content: space-between;
-    line-height: 24px;
-    text-transform: none;
-  }
-  svg {
-    align-self: flex-end;
-    height: auto;
-    max-height: 45px;
-    width: 42px;
-  }
-`;
 
 const HomeButtons = () => (
   <>
